+++ pkg/collector/corechecks/gpu/spec/architectures.yaml
